Revert "Bump pybind11"
This reverts commit 53672c156e9ab6398f18f72efb967d97d0f20287.

Running regression test in #11 isolated a difference in output caused by
the upgrade of pybind11 from 2.9.1 to 2.12.0. Further tests on pybind11
versions showed the change appeared between 2.10.0 to 2.10.1. A large
feature in 2.10.1 is GIL changes, but we need more tests to confirm if
this is the cause. For the time being, rolling back to 2.9.1 until we
have specific tests in place